requirements.txt
scapy
colorama
tabulate
fpdf2
